게시글 삭제 제한

+++ src/main/java/com/springbook/view/board/BoardController.java
@@ -6,6 +6,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -36,11 +38,11 @@ public class BoardController {
 		return boardListVO;
 	}
 
-	// ±Û µî·Ï
+	// ï¿½ï¿½ ï¿½ï¿½ï¿½
 	@RequestMapping(value = "/insertBoard.do")
 	public String insertBoard(BoardVO vo) throws IOException {
-		
-		// ÆÄÀÏ ¾÷·Îµå Ã³¸®
+		System.out.println("ì¶œë ¥ : " + vo.getTitle() + vo.getWriter() + vo.getContent());
+		// ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½Îµï¿½ Ã³ï¿½ï¿½
 		MultipartFile uploadFile = vo.getUploadFile();
 		if (!uploadFile.isEmpty()) {
 			String fileName = uploadFile.getOriginalFilename();
@@ -50,37 +52,44 @@ public class BoardController {
 		return "getBoardList.do";
 	}
 
-	// ±Û ¼öÁ¤
+	// ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
 	@RequestMapping("/updateBoard.do")
 	public String updateBoard(@ModelAttribute("board") BoardVO vo) {
 		boardService.updateBoard(vo);
 		return "getBoardList.do";
 	}
 
-	// ±Û »èÁ¦
+	// ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
 	@RequestMapping("/deleteBoard.do")
-	public String deleteBoard(BoardVO vo) {
-		boardService.deleteBoard(vo);
-		return "getBoardList.do";
+	public String deleteBoard(BoardVO vo, HttpSession session) {
+		System.out.println("ì‘ì„±ì : " + vo.getWriter()+"seq : "+vo.getSeq()+"íƒ€ì´í‹€ : "+vo.getTitle()+ "ì„¸ì…˜ : "+session.getAttribute("userName"));
+
+	if(session.getAttribute("userName").equals(vo.getWriter())) {
+			boardService.deleteBoard(vo);
+			return "getBoardList.do";
+	}else {
+		return "deleteBoardError.jsp";
+	}
+	
 	}
 
-	// ±Û »ó¼¼ Á¶È¸
+	// ï¿½ï¿½ ï¿½ï¿½ ï¿½ï¿½È¸
 	@RequestMapping("/getBoard.do")
 	public String getBoard(BoardVO vo, Model model) {
-		model.addAttribute("board", boardService.getBoard(vo)); // Model Á¤º¸ ÀúÀå
-		return "getBoard.jsp"; // View ÀÌ¸§ ¸®ÅÏ
+		model.addAttribute("board", boardService.getBoard(vo)); // Model ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
+		return "getBoard.jsp"; // View ï¿½Ì¸ï¿½ ï¿½ï¿½ï¿½ï¿½
 	}
 
-	// °Ë»ö Á¶°Ç ¸ñ·Ï ¼³Á¤
+	// ï¿½Ë»ï¿½ ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
 	@ModelAttribute("conditionMap")
 	public Map<String, String> searchConditionMap() {
 		Map<String, String> conditionMap = new HashMap<String, String>();
-		conditionMap.put("Á¦¸ñ", "TITLE");
-		conditionMap.put("³»¿ë", "CONTENT");
+		conditionMap.put("ï¿½ï¿½ï¿½ï¿½", "TITLE");
+		conditionMap.put("ï¿½ï¿½ï¿½ï¿½", "CONTENT");
 		return conditionMap;
 	}
 
-	// ±Û ¸ñ·Ï °Ë»ö
+	// ï¿½ï¿½ ï¿½ï¿½ï¿½ ï¿½Ë»ï¿½
 	@RequestMapping("/getBoardList.do")
 	public String getBoardList(BoardVO vo, Model model) {
 		// Null Check
@@ -88,8 +97,8 @@ public class BoardController {
 			vo.setSearchCondition("TITLE");
 		if (vo.getSearchKeyword() == null)
 			vo.setSearchKeyword("");
-		// Model Á¤º¸ ÀúÀå
+		// Model ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
 		model.addAttribute("boardList", boardService.getBoardList(vo));
-		return "getBoardList.jsp"; // View ÀÌ¸§ ¸®ÅÏ
+		return "getBoardList.jsp"; // View ï¿½Ì¸ï¿½ ï¿½ï¿½ï¿½ï¿½
 	}
 }
